extensions/when_evaluators/calendar_rrule/requirements.txt
-e ../lib
tzdata>=2023.3
python-dateutil>=2.8.2,<3
